Implementation Include to show CategoryName in Product Table View

+++ BulkyWeb/Areas/Admin/Controllers/ProductController.cs
@@ -20,13 +20,13 @@ public class ProductController : Controller
 
     public ActionResult Index()
     {
-        var objProductList = _unitOfWork.Product.GetAll().ToList();
+        var objProductList = _unitOfWork.Product.GetAll("Category").ToList();
         return View(objProductList);
     }
 
     public IActionResult Upsert(int? id)
     {
-        IEnumerable<SelectListItem> CategoryList = _unitOfWork.Category.GetAll().Select(u => new SelectListItem
+        var CategoryList = _unitOfWork.Category.GetAll().Select(u => new SelectListItem
         {
             Text = u.Name,
             Value = u.Id.ToString()
@@ -38,17 +38,11 @@ public class ProductController : Controller
             CategoryList = CategoryList,
             Product = new Product()
         };
-        if (id == null || id == 0)
-        {
-            return View(productVm);
-        }
-        else
-        {
-            //update
-            productVm.Product = _unitOfWork.Product.Get(u => u.Id == id);
-            return View(productVm);
-        }
-        
+        if (id == null || id == 0) return View(productVm);
+
+        //update
+        productVm.Product = _unitOfWork.Product.Get(u => u.Id == id);
+        return View(productVm);
     }
 
     [HttpPost]
@@ -65,43 +59,37 @@ public class ProductController : Controller
             });
             return View(productVM);
         }
-        
-        string wwwRootPath = _webHostEnvironment.WebRootPath;
+
+        var wwwRootPath = _webHostEnvironment.WebRootPath;
         if (file != null)
         {
-            string fileName = Guid.NewGuid().ToString() + Path.GetExtension(file.FileName);
-            string productPath = Path.Combine(wwwRootPath, @"images\product");
+            var fileName = Guid.NewGuid() + Path.GetExtension(file.FileName);
+            var productPath = Path.Combine(wwwRootPath, @"images\product");
 
             if (!string.IsNullOrEmpty(productVM.Product.ImageUrl))
             {
                 //delete the old image
                 var oldImagePath = Path.Combine(wwwRootPath, productVM.Product.ImageUrl.TrimStart('\\'));
-                if (System.IO.File.Exists(oldImagePath))
-                {
-                    System.IO.File.Delete(oldImagePath);
-                }
+                if (System.IO.File.Exists(oldImagePath)) System.IO.File.Delete(oldImagePath);
             }
-            
-            using (var fileStream = new FileStream(Path.Combine(productPath,fileName),FileMode.Create))
+
+            using (var fileStream = new FileStream(Path.Combine(productPath, fileName), FileMode.Create))
             {
                 file.CopyTo(fileStream);
             }
+
             productVM.Product.ImageUrl = @"\images\product\" + fileName;
         }
         else
         {
-            productVM.Product.ImageUrl = "";
+            if (productVM.Product.Id == 0) productVM.Product.ImageUrl = "";
         }
 
         if (productVM.Product.Id == 0)
-        {
             _unitOfWork.Product.Add(productVM.Product);
-        }
         else
-        {
             _unitOfWork.Product.Update(productVM.Product);
-        }
-        
+
         _unitOfWork.Save();
         TempData["success"] = "Product created successfully";
         return RedirectToAction("Index");
